# swagger.yaml
openapi: 3.0.0
paths:
  /api/v1/auth/passwordless-login:
    post:
      operationId: AuthController_passwordlessLogin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Request login code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
      tags: &a1
        - Authentication
  /api/v1/auth/passwordless-token:
    post:
      operationId: AuthController_passwordlessToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthTokenRequest"
      responses:
        "200":
          description: Request access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
      tags: *a1
  /api/v1/auth/signup:
    post:
      operationId: AuthController_signup
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: signup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
      tags: *a1
  /api/v1/auth/register/uploadDocument:
    put:
      operationId: AuthController_updateUserkycidcard
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                doc:
                  type: string
                  format: binary
                docType:
                  type: string
                  enum:
                    - idCardFront
                    - idCardBack
                  description: "ENUM: idCardFront, idCardBack"
              required:
                - doc
                - docType
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      summary: Cargar documentos KYC
      tags: *a1
  /api/v1/auth/confirmDetails:
    put:
      operationId: AuthController_confirmDetails
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmDetailsDto"
      responses:
        "200":
          description: User details updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
      security:
        - bearer: []
      summary: Confirm user personal details (KYC)
      tags: *a1
  /api/v1/auth/updateAddress:
    put:
      operationId: AuthController_updateAddress
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAddressDto"
      responses:
        "200":
          description: Address updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: "Forbidden: update failed"
      security:
        - bearer: []
      summary: Update user address
      tags: *a1
  /api/v1/backoffice/users/active:
    get:
      operationId: BackOfficeController_getActiveUsers
      parameters: []
      responses:
        "200":
          description: ""
      security: &a2
        - bearer: []
      tags: &a3
        - backoffice
  /api/v1/backoffice/users/kyc-rejected:
    get:
      operationId: BackOfficeController_getUsersWithRejectedKYC
      parameters: []
      responses:
        "200":
          description: ""
      security: *a2
      tags: *a3
  /api/v1/backoffice/wallet/balance:
    get:
      operationId: BackOfficeController_getCollectingWalletBalance
      parameters: []
      responses:
        "200":
          description: ""
      security: *a2
      tags: *a3
  /api/v1/backoffice/account/{userId}:
    get:
      operationId: BackOfficeController_getUserAccount
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      security: *a2
      tags: *a3
  /api/v1/backoffice/transactions:
    get:
      operationId: BackOfficeController_getAllTransactions
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
      responses:
        "200":
          description: ""
      security: *a2
      tags: *a3
info:
  title: Backoffice API Docs
  description: ""
  version: "0.1"
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
          example: name@company.domain
      required:
        - email
    AuthLoginResponse:
      type: object
      properties:
        code:
          type: string
          example: some-token-request
      required:
        - code
    AuthTokenRequest:
      type: object
      properties:
        code:
          type: string
          example: some-token-request
        token:
          type: string
          example: "1111"
      required:
        - code
        - token
    AuthTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: jwt-token-format
      required:
        - accessToken
    ConfirmDetailsDto:
      type: object
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        gender:
          type: string
          example: male
          description: Gender identity of the user
        birthdate:
          type: string
          example: 1990-01-01
          description: Birthdate in YYYY-MM-DD format
      required:
        - firstName
        - lastName
        - gender
        - birthdate
    User:
      type: object
      properties: {}
    UpdateAddressDto:
      type: object
      properties:
        street:
          type: string
          example: 123 Main St
        city:
          type: string
          example: Springfield
        state:
          type: string
          example: Illinois
        country:
          type: string
          example: US
        postalCode:
          type: string
          example: "62704"
      required:
        - street
        - city
        - state
        - country
        - postalCode